renamed vault -> vaultkv and corrections in KAP accordingly
